First draft migration

+++ crates/cubek-attention/src/components/global/simple/setup.rs
@@ -54,7 +54,7 @@ impl<
         let plane_role_config = PlaneRoleConfig::new_unspecialized(stage_config.num_planes());
 
         let query_gmem_config = GlobalMemoryConfig {
-            line_size: blueprint.line_sizes.query as u32,
+            line_size: blueprint.line_sizes.query,
             check_row_bounds: blueprint.check_bounds.seq_q,
             check_col_bounds: blueprint.check_bounds.head_dim,
             matrix_layout: MatrixLayout::RowMajor,
@@ -62,7 +62,7 @@ impl<
         };
 
         let mask_gmem_config = GlobalMemoryConfig {
-            line_size: blueprint.line_sizes.mask as u32,
+            line_size: blueprint.line_sizes.mask,
             check_row_bounds: blueprint.check_bounds.seq_q,
             check_col_bounds: blueprint.check_bounds.seq_kv,
             matrix_layout: MatrixLayout::RowMajor,
@@ -70,7 +70,7 @@ impl<
         };
 
         let key_gmem_config = GlobalMemoryConfig {
-            line_size: blueprint.line_sizes.key as u32,
+            line_size: blueprint.line_sizes.key,
             check_row_bounds: blueprint.check_bounds.seq_kv,
             check_col_bounds: blueprint.check_bounds.head_dim,
             matrix_layout: MatrixLayout::RowMajor,
@@ -78,7 +78,7 @@ impl<
         };
 
         let value_gmem_config = GlobalMemoryConfig {
-            line_size: blueprint.line_sizes.value as u32,
+            line_size: blueprint.line_sizes.value,
             check_row_bounds: blueprint.check_bounds.seq_kv,
             check_col_bounds: blueprint.check_bounds.val_dim,
             matrix_layout: MatrixLayout::RowMajor,
@@ -86,7 +86,7 @@ impl<
         };
 
         let out_gmem_config = GlobalMemoryConfig {
-            line_size: blueprint.line_sizes.out as u32,
+            line_size: blueprint.line_sizes.out,
             check_row_bounds: blueprint.check_bounds.seq_q,
             check_col_bounds: blueprint.check_bounds.val_dim,
             matrix_layout: MatrixLayout::RowMajor,
