+++ crates/cubek-reduce/src/components/global/cube.rs
@@ -21,7 +21,7 @@ impl GlobalFullCubeReduce {
     pub fn execute<P: ReducePrecision, Out: Numeric, I: ReduceInstruction<P>>(
         input: &VirtualTensor<P::EI>,
         output: &mut VirtualTensor<Out, ReadWrite>,
-        reduce_axis: u32,
+        reduce_axis: usize,
         inst: &I,
         #[comptime] line_mode: LineMode,
         #[comptime] blueprint: CubeBlueprint,
@@ -94,16 +94,16 @@ impl GlobalFullCubeReduce {
 
         #[allow(clippy::collapsible_if)]
         if comptime!(commit_required) {
-            if worker_pos == 0u32 {
+            if worker_pos == 0 {
                 writer.commit();
             }
         }
     }
 
-    fn worker_pos(#[comptime] blueprint: CubeBlueprint) -> u32 {
+    fn worker_pos(#[comptime] blueprint: CubeBlueprint) -> usize {
         match comptime!(blueprint.use_planes) {
-            true => UNIT_POS_Y,
-            false => UNIT_POS,
+            true => UNIT_POS_Y as usize,
+            false => UNIT_POS as usize,
         }
     }
 
@@ -111,8 +111,8 @@ impl GlobalFullCubeReduce {
     fn reduce_shared<P: ReducePrecision, Out: Numeric, I: ReduceInstruction<P>>(
         input: &VirtualTensor<P::EI>,
         output: &mut VirtualTensor<Out, ReadWrite>,
-        reduce_axis: u32,
-        reduce_index: u32,
+        reduce_axis: usize,
+        reduce_index: usize,
         inst: &I,
         idle: CubeOption<bool>,
         #[comptime] line_mode: LineMode,
@@ -173,7 +173,7 @@ fn reduce_scan<P: ReducePrecision, I: ReduceInstruction<P>>(
     inst: &I,
     accumulator: &mut I::SharedAccumulator,
     result: &mut I::AccumulatorItem,
-    #[comptime] size: u32,
+    #[comptime] size: usize,
 ) {
     for i in 0..size {
         let item = I::SharedAccumulator::read(accumulator, i);
@@ -211,8 +211,8 @@ fn reduce_tree<P: ReducePrecision, I: ReduceInstruction<P>>(
     inst: &I,
     accumulator: &mut I::SharedAccumulator,
     result: &mut I::AccumulatorItem,
-    worker_index: u32,
-    #[comptime] size: u32,
+    worker_index: usize,
+    #[comptime] size: usize,
 ) {
     if comptime!(size.is_power_of_two()) {
         let mut num_active_units = size.runtime();
