+++ crates/cubek-reduce/src/components/global/plane.rs
@@ -21,12 +21,12 @@ impl GlobalFullPlaneReduce {
     pub fn execute<P: ReducePrecision, Out: Numeric, I: ReduceInstruction<P>>(
         input: &VirtualTensor<P::EI>,
         output: &mut VirtualTensor<Out, ReadWrite>,
-        reduce_axis: u32,
+        reduce_axis: usize,
         inst: &I,
         #[comptime] line_mode: LineMode,
         #[comptime] blueprint: PlaneReduceBlueprint,
     ) {
-        let write_index = CUBE_POS * CUBE_DIM_Y + UNIT_POS_Y;
+        let write_index = CUBE_POS * CUBE_DIM_Y as usize + UNIT_POS_Y as usize;
 
         let mut writer =
             Writer::<Out>::new::<P>(input, output, reduce_axis, write_index, line_mode);
@@ -74,8 +74,8 @@ impl GlobalFullPlaneReduce {
     fn reduce_single<P: ReducePrecision, Out: Numeric, I: ReduceInstruction<P>>(
         input: &VirtualTensor<P::EI>,
         output: &mut VirtualTensor<Out, ReadWrite>,
-        reduce_axis: u32,
-        reduce_index: u32,
+        reduce_axis: usize,
+        reduce_index: usize,
         inst: &I,
         idle: CubeOption<bool>,
         #[comptime] line_mode: LineMode,
